Move web vitals definitions to separate file
- Create lib/web-vitals.ts with all 5 web vitals definitions:
  - webVitalsEvents (internal pipe)
  - webVitalsCurrent
  - webVitalsDistribution
  - webVitalsRoutes
  - webVitalsTimeseries
- Remove web vitals from lib/endpoints.ts
- Remove lib/client.ts (SDK includes directly from tinybird.json)
- Update tinybird.json to include all 5 TypeScript files
